% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{treestack-package}
\alias{treestack-package}
\title{Stacked Treemap}
\description{
\tabular{ll}{
Package: \tab treestack\cr
Type: \tab Package\cr
License: \tab GPL-(>=2)\cr
LazyLoad: \tab yes\cr
}
}
\details{
Visualize data iteratively aggregated by categories using stacked barplots.  The color and size of the bars can be determined by continuous variables.
}
\author{
Noah Peart \email{noah.v.peart@gmail.com}
}
\keyword{graph}
\keyword{tree}
\keyword{treemap}

