% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treestack.R
\docType{package}
\name{treestack}
\alias{treestack}
\alias{treestack-package}
\title{reestack: a package to make stacked bar version of treemaps}
\description{
reestack: a package to make stacked bar version of treemaps
}

